plugins/honcho-bridge/requirements.txt
# Official Honcho SDK
honcho-ai>=0.6.0

# Ollama client (for local LLM)
ollama>=0.1.0

# Postgres driver (for Honcho database)
psycopg2-binary>=2.9.0

# YAML parsing (for wiki import/export)
pyyaml>=6.0
